# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_mwork_mwork-api_internal_pkg_response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  github_com_mwork_mwork-api_internal_pkg_response.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  github_com_mwork_mwork-api_internal_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Meta'
      success:
        type: boolean
    type: object
...
  internal_domain_admin.GrantCreditsRequest:
    properties:
      amount:
        maximum: 1000000
        minimum: 1
        type: integer
      reason:
        maxLength: 500
        minLength: 3
        type: string
    required:
    - amount
    - reason
    type: object
...
  internal_domain_admin.ResolveRequest:
    properties:
      action:
        enum:
        - warn
        - suspend
        - delete
        - dismiss
        type: string
      notes:
        maxLength: 500
        type: string
    required:
    - action
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API server for model agency aggregator.
  termsOfService: http://swagger.io/terms/
  title: MWork API
  version: "1.0"
paths:
...
  /admin/reports:
    get:
      parameters:
      - description: Лимит
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      - description: Статус (pending|resolved|rejected)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Все жалобы (админ)
      tags:
      - Moderation Admin
...
  /admin/reports/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID репорта
        in: path
        name: id
        required: true
        type: string
      - description: Решение по репорту
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_domain_admin.ResolveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Обработать репорт
      tags:
      - Admin Reports
...
  /admin/users/{id}/credits:
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Баланс кредитов пользователя
      tags:
      - Admin Credits
  /admin/users/{id}/credits/grant:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      - description: Данные начисления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_domain_admin.GrantCreditsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Начислить кредиты пользователю
      tags:
      - Admin Credits
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"